Rename transitions directory
Captialize to match Architect so we can share transitions for now.
