+++ src/renderer/components/Transitions/Fade.js
