+++ src/renderer/components/Transitions/Modal.js
