+++ src/renderer/components/Transitions/index.js
